move cache history into the repositoryImp instead of the domain layer

+++ domain/vod/src/main/java/usecase/RemoveSearchHistoryItemUseCase.kt
@@ -5,7 +5,7 @@ import repository.SearchHistoryRepository
 class RemoveSearchHistoryItemUseCase(
     private val historyRepo: SearchHistoryRepository
 ) {
-    suspend fun execute(id: Long) {
+    suspend fun execute(id: Int) {
         historyRepo.removeSearchHistoryItem(id)
     }
 }
